requirements.txt
Flask
reportlab
python-docx
gunicorn
